feat: prevent dirs expanding with rm command

+++ lgsm/functions/alert.sh
@@ -12,7 +12,7 @@ fn_alert_log(){
 	info_config.sh
 	info_messages.sh
 	if [ -f "${alertlog}" ]; then
-		rm "${alertlog}"
+		rm -f "${alertlog:?}"
 	fi
 
 	{
