+++ lgsm/functions/command_dev_debug.sh
@@ -9,7 +9,7 @@ local commandaction="Dev-Debug"
 local function_selfname=$(basename "$(readlink -f "${BASH_SOURCE[0]}")")
 
 if [ -f "${rootdir}/.dev-debug" ]; then
-	rm "${rootdir}/.dev-debug"
+	rm "${rootdir:?}/.dev-debug"
 	fn_print_ok_nl "Disabled dev-debug"
 	fn_script_log_info "Disabled dev-debug"
 else
