+++ lgsm/functions/command_stop.sh
@@ -242,7 +242,7 @@ info_config.sh
 fn_stop_pre_check
 # Remove lockfile.
 if [ -f "${rootdir}/${lockselfname}" ]; then
-	rm -f "${rootdir}/${lockselfname}"
+	rm -f "${rootdir:?}/${lockselfname}"
 fi
 
 if [ -z "${exitbypass}" ]; then
